+++ examples/quick-start/nuxt.config.ts
@@ -1,7 +1,5 @@
-import type { ElectronOptions } from 'nuxt-electron'
-
 export default defineNuxtConfig({
   modules: [
-    ['nuxt-electron', <ElectronOptions>{}],
+    'nuxt-electron',
   ],
 })